+++ app/src/main/java/com/wepli/app/di/module/PendingIntentModule.kt
@@ -1,19 +1,20 @@
-package com.wepli.app.di
+package com.wepli.app.di.module
 
 import android.content.Context
+import android.content.Intent
+import com.wepli.app.MainActivity
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import util.notification.NotificationManager
 
 @Module
 @InstallIn(SingletonComponent::class)
-object UtilModule {
+object PendingIntentModule {
 
     @Provides
-    fun provideNotificationManager(@ApplicationContext context: Context): NotificationManager {
-        return NotificationManager(context)
+    fun bindMainActivityIntent(@ApplicationContext context: Context): Intent {
+        return Intent(context, MainActivity::class.java)
     }
 }
