+++ app/src/main/java/com/wepli/app/di/module/UtilModule.kt
@@ -1,20 +1,19 @@
-package com.wepli.app.di
+package com.wepli.app.di.module
 
 import android.content.Context
-import android.content.Intent
-import com.wepli.app.MainActivity
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
+import util.notification.NotificationManager
 
 @Module
 @InstallIn(SingletonComponent::class)
-object PendingIntentModule {
+object UtilModule {
 
     @Provides
-    fun bindMainActivityIntent(@ApplicationContext context: Context): Intent {
-        return Intent(context, MainActivity::class.java)
+    fun provideNotificationManager(@ApplicationContext context: Context): NotificationManager {
+        return NotificationManager(context)
     }
 }
